# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel API",
        "description": "Laravel API Documentation",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8097/api",
            "description": "Local server"
        },
        {
            "url": "https://api.prod.example.com/api",
            "description": "Production server"
        },
        {
            "url": "https://api.staging.example.com/api",
            "description": "Staging server"
        }
    ],
    "paths": {
...
        "/documents": {
            "get": {
                "tags": [
                    "Documents"
                ],
                "summary": "Get a list of documents",
                "operationId": "58aa02e52615eb94196343283d7e4b56",
                "responses": {
                    "200": {
                        "description": "Documents retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Document"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No documents found"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Documents"
                ],
                "summary": "Upload a new document",
                "operationId": "1376f0c660d7ac76c951592057ca0d19",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/DocumentRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Document uploaded successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Document"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Document could not be uploaded"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/documents/{uuid}": {
            "get": {
                "tags": [
                    "Documents"
                ],
                "summary": "Get a specific document",
                "operationId": "57dd60228c625f36e6c6907a984208f1",
                "parameters": [
                    {
                        "name": "uuid",
                        "in": "path",
                        "description": "The ID of the document",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Document retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Document"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Document not found"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Documents"
                ],
                "summary": "Delete a document",
                "operationId": "2c70c2f0dfdf713a529883c6f3ea69d7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The ID of the document",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Document deleted successfully"
                    },
                    "404": {
                        "description": "Document not found"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/documents/{uuid}/status": {
            "patch": {
                "tags": [
                    "Documents"
                ],
                "summary": "Update the status of a document",
                "operationId": "4b0b0a8b0f81a79783610bf65e3ee226",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The ID of the document",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "status": {
                                        "description": "The new status of the document",
                                        "type": "string",
                                        "enum": [
                                            "pending",
                                            "signed"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Document status updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Document"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Document not found"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/documents/{uuid}/download": {
            "get": {
                "tags": [
                    "Documents"
                ],
                "summary": "Download a document",
                "operationId": "3746ca3aa82a33a0c2f617b3f0b918af",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The ID of the document",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Document downloaded successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Document"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Document not found"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/documents/{uuid}/send-signature-request": {
            "post": {
                "tags": [
                    "Signature Requests"
                ],
                "summary": "Send a signature request for a document",
                "operationId": "bd03a27bd4bc7898e38422df03d621fa",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The ID of the document",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SignatureRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Signature request sent successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SignatureRequest"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "You do not own this document"
                    },
                    "400": {
                        "description": "Signature request could not be sent"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
...
        "/documents/{uuid}/add-signature": {
            "post": {
                "tags": [
                    "Signature Requests"
                ],
                "summary": "Add a signature to a document",
                "operationId": "3064f8c9edb47579f46592d13af25508",
                "parameters": [
                    {
                        "name": "documentId",
                        "in": "path",
                        "description": "Document ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Signature"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Document signed successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Signature"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Not authorized"
                    },
                    "400": {
                        "description": "Signature error"
                    },
                    "404": {
                        "description": "Signature request not found"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "Document": {
                "title": "Document",
                "description": "Document object",
                "required": [
                    "id",
                    "name",
                    "url"
                ],
                "properties": {
                    "id": {
                        "description": "The document ID",
                        "type": "integer",
                        "format": "int64"
                    },
                    "name": {
                        "description": "The document name",
                        "type": "string"
                    },
                    "filePath": {
                        "description": "The document URL",
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "DocumentRequest": {
                "title": "Document Request",
                "description": "Request body for document operations",
                "required": [
                    "document"
                ],
                "properties": {
                    "document": {
                        "description": "The document file to be uploaded",
                        "type": "string",
                        "format": "binary"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "passport": {
                "type": "oauth2",
                "description": "Laravel passport oauth2 security.",
                "in": "header",
                "scheme": "https",
                "flows": {
                    "password": {
                        "authorizationUrl": "http://lureg.test/oauth/authorize",
                        "tokenUrl": "http://lureg.test/oauth/token",
                        "refreshUrl": "http://lureg.test/token/refresh",
                        "scopes": []
                    }
                }
            }
        }
    },
...
}